# api/openapi.yaml
openapi: 3.1.0
info:
  title: agent-manager API
  version: 2.0.0-poc
  description: |
    Agent lifecycle for the Giant Swarm Agent Platform on kagent main
    (`kagent.dev/v1alpha3`).

    An agent is an `AgentTemplate` in the kagent namespace — description,
    system prompt, model config, skills pinned to immutable sources, one MCP
    tool binding — that kagent compiles for every `Harness` whose admission
    selector matches its labels (`kagent.dev/harness=<harness>`). The toolset
    the agent declares travels as its **toolset carrier**: a per-agent copy of
    the platform's `muster` `RemoteMCPServer` (`muster-<agent>`) carrying the
    `X-Muster-Toolset` header, which the template binds instead of the platform
    server. agent-manager composes both objects as the caller, validates what
    kagent main can express before anything is applied, and reads readiness
    from the template's per-Harness status. Conversations are `AgentInstance`s
    created from the compiled template over kagent's own gRPC API, not here.

    Clients MUST call `GET /api/v1/info` first: it reports the managed
    namespaces, the served kagent.dev API version, the platform's muster server
    and default Harness, the capability flags (a false flag names an argument
    an AgentTemplate cannot express — `perToolSelection`, `runtime`, `iconUrl`,
    `skillGitAuthSecret`, `mutableSkillRefs`) and `identity`: `caller` means
    the server is an OAuth 2.1 resource server — every request needs a bearer
    token the platform IdP issued (muster forwards the session's id_token, the
    portal sends the user's through the gateway) and every Kubernetes call is
    made as that user, so the user's RBAC decides and writes report
    `requestedBy`; `serviceAccount` means no OAuth, the service's own identity
    behind a trusted proxy. Without a valid token the OAuth layer answers `401`
    with a `WWW-Authenticate: Bearer` challenge (RFC 9728) before any handler
    runs.

    The same operations are available as MCP tools (streamable HTTP at `/mcp`,
    tool names in `x-mcp-tool` below; through muster they appear as
    `x_<server>_<tool>`).

    Ownership rules: a template applied by a Flux Kustomization (`managed:
    gitops`) or written by someone else (`managed: external`) is refused for
    update and delete unless `force`; so is an update of a template carrying
    fields agent-manager does not compose (`unmanagedFields`).
servers:
  - url: /
tags:
  - name: info
  - name: agents
  - name: modelconfigs
  - name: skills
  - name: health
paths:
  /healthz:
    get:
      tags: [health]
      summary: Liveness
      operationId: healthz
      responses:
        "200":
          description: alive
          content:
            text/plain:
              schema: { type: string, example: ok }
  /readyz:
    get:
      tags: [health]
      summary: Readiness (the API is serving)
      description: |
        Does not depend on the API server: the API stays reachable so clients
        read a failure from the response itself.
      operationId: readyz
      responses:
        "200":
          description: ready
          content:
            text/plain:
              schema: { type: string, example: ok }
  /api/v1/info:
    get:
      tags: [info]
      summary: Service version, kagent API versions, managed namespaces and capability flags
      operationId: getInfo
      x-mcp-tool: get_info
      responses:
        "200":
          description: the installation's identity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Info" }
  /api/v1/openapi.yaml:
    get:
      tags: [info]
      summary: This document
      operationId: getOpenAPI
      responses:
        "200":
          description: the contract
          content:
            application/yaml:
              schema: { type: string }
  /api/v1/agents:
    get:
      tags: [agents]
      summary: List the agents of a namespace
      description: |
        Every AgentTemplate of the namespace with its toolset carrier when it
        has one: the declaration, the toolset (or `implicitFullAccess`), the
        per-Harness status and who manages it.
      operationId: listAgents
      x-mcp-tool: list_agents
      parameters:
        - $ref: "#/components/parameters/namespaceQuery"
      responses:
        "200":
          description: agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/Error" }
    post:
      tags: [agents]
      summary: Create an agent
      description: |
        Composes the toolset carrier and the AgentTemplate, checks the name,
        the toolset (required), the modelConfig against the namespace, the
        Harness admission and the skill sources, then applies the carrier and
        the template (a template that fails to apply takes its carrier with
        it). A check failure writes nothing (`400 invalid_request`, the
        message names the valid model configs / the Harnesses and what they
        admit). `409 conflict` when an AgentTemplate or a RemoteMCPServer of
        the carrier's name exists.
      operationId: createAgent
      x-mcp-tool: create_agent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Spec" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateResult" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/agents/validate:
    post:
      tags: [agents]
      summary: Dry run of create (or update)
      description: |
        Composes and validates without writing. With `update: true` the
        non-empty fields of the body are validated as a change to the existing
        agent's declaration.
      operationId: validateAgent
      x-mcp-tool: validate_agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Spec"
                - type: object
                  properties:
                    update: { type: boolean, description: Validate as an update of the existing agent }
                    force: { type: boolean, description: With update, ignore the ownership guards }
      responses:
        "200":
          description: the verdict (valid may be false)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidateResult" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/agents/{namespace}/{name}:
    parameters:
      - $ref: "#/components/parameters/namespacePath"
      - $ref: "#/components/parameters/namePath"
    get:
      tags: [agents]
      summary: Get one agent
      operationId: getAgent
      x-mcp-tool: get_agent
      responses:
        "200":
          description: the agent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [agents]
      summary: Update an agent
      description: |
        Merges the given fields into the agent's declaration (absent fields
        stay, an empty string clears a field, `skills` replaces its whole
        block, `toolset` replaces the whole list — the way to assign a toolset
        to an agent that reports `implicitFullAccess`), re-composes and writes
        what differs: a toolset change updates the carrier only; anything else
        updates the AgentTemplate, which makes kagent compile a **new
        revision** for every admitting Harness while running AgentInstances
        keep the revision they were created with (`note` says so). `409
        conflict` for GitOps-owned or externally written templates and for
        templates carrying `unmanagedFields` unless `force`. `400` for a body
        still carrying the removed `toolNames`, `runtime` or `iconUrl`.
      operationId: updateAgent
      x-mcp-tool: update_agent
      parameters:
        - name: force
          in: query
          schema: { type: boolean }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Update" }
      responses:
        "200":
          description: before / after
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpdateResult" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [agents]
      summary: Delete an agent
      description: |
        Deletes the AgentTemplate and its toolset carrier (only when
        agent-manager created the carrier; `toolsetCarrierKept` says why one
        stays). GitOps-owned and externally written templates need `force`.
      operationId: deleteAgent
      x-mcp-tool: delete_agent
      parameters:
        - name: force
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: what was removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/agents/{namespace}/{name}/status:
    parameters:
      - $ref: "#/components/parameters/namespacePath"
      - $ref: "#/components/parameters/namePath"
    get:
      tags: [agents]
      summary: Compact status verdict
      description: |
        The AgentTemplate's per-Harness status (Accepted, ResolvedRefs,
        Compatible, Ready, revisions, compile warnings) and the toolset
        carrier's acceptance, folded into `verdict` (ready | progressing |
        failed | unknown) and `summary`. `ready` = any admitting Harness
        reports Ready for the current generation.
      operationId: getAgentStatus
      x-mcp-tool: get_agent_status
      responses:
        "200":
          description: the status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/modelconfigs:
    get:
      tags: [modelconfigs]
      summary: List the kagent ModelConfigs of a namespace
      operationId: listModelConfigs
      x-mcp-tool: list_model_configs
      parameters:
        - $ref: "#/components/parameters/namespaceQuery"
      responses:
        "200":
          description: model configs
          content:
            application/json:
              schema:
                type: object
                properties:
                  modelConfigs:
                    type: array
                    items: { $ref: "#/components/schemas/ModelConfig" }
  /api/v1/skills:
    get:
      tags: [skills]
      summary: Discover skills in the configured repositories
      description: |
        Every `SKILL.md` of a GitHub repository is one skill (frontmatter
        `name`/`description`), read at the commit the branch resolves to and
        returned with that `commit` — the `gitRefs` entry an agent mounts
        (`ref` = the commit). Cached briefly; `refresh=true` re-reads GitHub.
        `501 unsupported` when no repositories are configured.
      operationId: listSkills
      x-mcp-tool: list_skills
      parameters:
        - name: repository
          in: query
          description: Only this repository (https://github.com/<owner>/<repo>)
          schema: { type: string }
        - name: ref
          in: query
          description: Git ref; default the default branch
          schema: { type: string }
        - name: refresh
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: skills
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillsResult" }
        "400": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
components:
  parameters:
    namespaceQuery:
      name: namespace
      in: query
      description: Namespace; default the installation's kagent namespace
      schema: { type: string }
    namespacePath:
      name: namespace
      in: path
      required: true
      schema: { type: string }
    namePath:
      name: name
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [not_found, invalid_request, conflict, forbidden, unauthenticated, unsupported, backend_error]
            message: { type: string }
    Info:
      type: object
      properties:
        version: { type: string }
        namespaces:
          type: object
          properties:
            default: { type: string }
            managed: { type: array, items: { type: string } }
        capabilities:
          type: object
          additionalProperties: { type: boolean }
          description: >-
            list, get, create, update, delete, status, validate, modelConfigs, skills, writesAsCaller (true when
            identity is caller), harnesses, toolsetCarrier; and what an AgentTemplate cannot express, always false
            on kagent main: perToolSelection (tools[] needs controller-side discovery), runtime (the Harness is
            the runtime), iconUrl, skillGitAuthSecret (skill sources are read anonymously), mutableSkillRefs
            (git skills pin a commit, OCI skills a digest).
        identity:
          type: string
          enum: [caller, serviceAccount]
          description: >-
            How calls reach the Kubernetes API. `caller`: the server runs as an OAuth 2.1 resource
            server (mcp-oauth) with downstream OAuth — every call presents the signed-in user's IdP
            id_token (forwarded by muster, or sent by the portal through the gateway) and the user's
            RBAC governs; the ServiceAccount holds no permissions. `serviceAccount`: no OAuth, the
            service's own identity behind a trusted proxy.
        apiVersions:
          type: object
          description: The served kagent.dev version, discovered from the API server (agenttemplates).
          properties:
            agentTemplate: { type: string, example: kagent.dev/v1alpha3 }
            remoteMcpServer: { type: string, example: kagent.dev/v1alpha3 }
            harness: { type: string, example: kagent.dev/v1alpha3 }
            modelConfig: { type: string, example: kagent.dev/v1alpha3 }
        kagent:
          type: object
          properties:
            musterServer: { type: string, description: The platform RemoteMCPServer every carrier is copied from (muster) }
            defaultHarness: { type: string, description: The Harness an agent runs on when the request names none (kagent) }
            toolsetHeader: { type: string, example: X-Muster-Toolset }
        skillsRepositories: { type: array, items: { type: string } }
    SkillGitRef:
      type: object
      required: [url, ref]
      properties:
        url: { type: string, description: Repository URL (http or https) }
        path: { type: string, description: Skill directory; empty for the repository root }
        ref: { type: string, pattern: '^([0-9a-fA-F]{40}|[0-9a-fA-F]{64})$', description: Full git commit id; list_skills reports it as `commit` }
        name: { type: string, description: Skill name under the template; defaults to the directory name }
    Skills:
      type: object
      properties:
        refs: { type: array, items: { type: string, pattern: '^[^\s@]+@sha256:[0-9a-f]{64}$' }, description: Digest-pinned OCI skill references }
        gitRefs: { type: array, items: { $ref: "#/components/schemas/SkillGitRef" } }
    Toolset:
      type: array
      minItems: 1
      maxItems: 32
      items: { type: string, pattern: '^(preset|server|workflow|tool):[^\s,]+$' }
      description: |
        The selectors that bound which of the gateway's tools the agent's
        meta-tools can see and call — carried as the `X-Muster-Toolset`
        header of the agent's toolset carrier (RemoteMCPServer
        `muster-<agent>`, a copy of the platform's `muster` server that the
        AgentTemplate binds), resolved by muster per caller. Each selector is
        `preset:<name>`, `server:<name>`, `workflow:<name>` or `tool:<name>`
        (exact names, at most 32 — define a preset for more). Shipped presets:
        `read-only`, `none` (a chat-only agent without tools), `infrastructure`,
        `agent-platform`, `full` (every tool the gateway exposes). `toolset:` is
        reserved, `label:` exists inside presets only, an empty list is refused
        (use `preset:none`). The former `toolNames` never narrowed anything
        against muster (kagent filters muster's meta-tools only) and is refused
        with that explanation.
    Spec:
      type: object
      required: [name, modelConfig, toolset]
      properties:
        namespace: { type: string }
        name: { type: string, description: DNS-1123 technical name (the AgentTemplate name), max 63 chars }
        displayName: { type: string, maxLength: 63, description: Annotation ui.giantswarm.io/display-name }
        description: { type: string, description: spec.description }
        systemMessage: { type: string, description: spec.systemPrompt }
        modelConfig: { type: string, description: An existing ModelConfig of the namespace }
        harness:
          type: string
          description: >-
            The kagent Harness that runs the agent (the runtime); the template is labelled
            kagent.dev/harness=<harness> and a Harness admitting it must exist. Empty: the
            installation default (info.kagent.defaultHarness). Replaces the former `runtime`.
        skills: { $ref: "#/components/schemas/Skills" }
        toolset: { $ref: "#/components/schemas/Toolset" }
        labels: { type: object, additionalProperties: { type: string }, description: Extra labels on the AgentTemplate (a Harness with a foreign selector is matched through them) }
        annotations: { type: object, additionalProperties: { type: string } }
    Update:
      type: object
      description: Partial change; absent fields stay, "" clears, skills replace their block, toolset replaces the whole list
      properties:
        displayName: { type: string }
        description: { type: string }
        systemMessage: { type: string }
        modelConfig: { type: string }
        harness: { type: string, description: Move the agent to another Harness; "" restores the default }
        skills: { $ref: "#/components/schemas/Skills" }
        toolset: { $ref: "#/components/schemas/Toolset" }
        labels: { type: object, additionalProperties: { type: string } }
        annotations: { type: object, additionalProperties: { type: string } }
        force: { type: boolean }
    Condition:
      type: object
      properties:
        type: { type: string }
        status: { type: string }
        reason: { type: string }
        message: { type: string }
        lastTransitionTime: { type: string }
        observedGeneration: { type: integer }
    HarnessStatus:
      type: object
      description: What kagent reports for the template on one admitting Harness (status.harnesses[])
      properties:
        harness: { type: string }
        ready: { type: [boolean, "null"] }
        accepted: { type: [boolean, "null"] }
        resolvedRefs: { type: [boolean, "null"] }
        compatible: { type: [boolean, "null"] }
        desiredRevision: { type: string }
        latestSuccessfulRevision: { type: string, description: Instances pin the revision they were created with }
        conditions: { type: array, items: { $ref: "#/components/schemas/Condition" } }
        warnings: { type: array, items: { type: string }, description: kagent's non-blocking compile decisions }
    ToolsetCarrier:
      type: object
      properties:
        name: { type: string, example: muster-sre }
        exists: { type: boolean }
        accepted: { type: [boolean, "null"] }
        message: { type: string }
        header: { type: string, description: The X-Muster-Toolset value the carrier sends }
    Agent:
      type: object
      properties:
        name: { type: string }
        namespace: { type: string }
        displayName: { type: string }
        description: { type: string }
        modelConfig: { type: string }
        systemMessage: { type: string }
        harness: { type: string, description: The kagent.dev/harness label; absent when the template carries none }
        skills: { $ref: "#/components/schemas/Skills" }
        toolset: { $ref: "#/components/schemas/Toolset", description: The declared toolset (the carrier's X-Muster-Toolset header); absent when none is declared }
        implicitFullAccess: { type: boolean, description: True when the template binds the platform's muster server directly (no carrier) — its tools are every tool the gateway exposes to the caller; assign a toolset with update_agent }
        toolsetCarrier: { $ref: "#/components/schemas/ToolsetCarrier" }
        ready: { type: [boolean, "null"], description: Any admitting Harness reports Ready; null while none has reported }
        harnesses: { type: array, items: { $ref: "#/components/schemas/HarnessStatus" } }
        managed: { type: string, enum: [agent-manager, gitops, external] }
        generation: { type: integer }
        observedGeneration: { type: integer }
        unmanagedFields: { type: array, items: { type: string }, description: Template fields agent-manager does not compose (bucket skills, plugins, prompt templates, agent tool bindings, other MCP servers); an update overwrites them only with force }
        spec: { type: object, description: The AgentTemplate spec as served (the kagent contract) }
    Manifests:
      type: object
      properties:
        agentTemplate: { type: string, description: YAML }
        toolsetCarrier: { type: string, description: YAML; empty when the template binds the platform server directly }
    ValidateResult:
      type: object
      properties:
        valid: { type: boolean }
        mode: { type: string, enum: [create, update] }
        errors: { type: array, items: { type: string } }
        manifests: { $ref: "#/components/schemas/Manifests" }
    CreateResult:
      type: object
      properties:
        agent: { $ref: "#/components/schemas/Agent" }
        manifests: { $ref: "#/components/schemas/Manifests" }
        created:
          type: object
          properties:
            agentTemplate: { type: boolean }
            toolsetCarrier: { type: boolean }
        status: { $ref: "#/components/schemas/Status" }
        requestedBy: { type: string, description: The authenticated caller the write ran as (email, else subject); absent without OAuth }
    UpdateResult:
      type: object
      properties:
        agent: { $ref: "#/components/schemas/Agent" }
        before: { type: object, description: The declaration before (displayName, description, systemMessage, modelConfig, harness, skills, toolset, labels, annotations) }
        after: { type: object }
        changed: { type: array, items: { type: string }, description: Dotted declaration paths that changed, e.g. description, toolset, skills.gitRefs, labels.tenant }
        note: { type: string, description: Present when the AgentTemplate changed — kagent compiles a new revision; running instances keep theirs }
        requestedBy: { type: string, description: The authenticated caller the write ran as }
        manifests: { $ref: "#/components/schemas/Manifests" }
    DeleteResult:
      type: object
      properties:
        name: { type: string }
        namespace: { type: string }
        agentTemplateDeleted: { type: boolean }
        toolsetCarrierDeleted: { type: boolean }
        toolsetCarrierKept: { type: string, description: Why a carrier of that name stays (not created by agent-manager) }
        requestedBy: { type: string, description: The authenticated caller the delete ran as }
    Status:
      type: object
      properties:
        name: { type: string }
        namespace: { type: string }
        verdict: { type: string, enum: [ready, progressing, failed, unknown] }
        summary: { type: string }
        template:
          type: object
          properties:
            exists: { type: boolean }
            generation: { type: integer }
            observedGeneration: { type: integer }
            harnesses: { type: array, items: { $ref: "#/components/schemas/HarnessStatus" } }
        toolsetCarrier: { $ref: "#/components/schemas/ToolsetCarrier" }
    ModelConfig:
      type: object
      properties:
        name: { type: string }
        namespace: { type: string }
        provider: { type: string }
        model: { type: string }
        accepted: { type: [boolean, "null"] }
        message: { type: string }
        managedBy: { type: string }
    Skill:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        repoUrl: { type: string }
        path: { type: string }
        ref: { type: string, description: The branch that was read }
        commit: { type: string, description: The commit the branch resolved to — what a template pins }
    SkillsResult:
      type: object
      properties:
        repositories:
          type: array
          items:
            type: object
            properties:
              repoUrl: { type: string }
              ref: { type: string }
              commit: { type: string }
              skills: { type: array, items: { $ref: "#/components/schemas/Skill" } }
              truncated: { type: boolean }
              error: { type: string }
              fetchedAt: { type: string, format: date-time }
        skills: { type: array, items: { $ref: "#/components/schemas/Skill" } }